requirements.txt
appnope==0.1.0
appscript==1.0.1
beautifulsoup4==4.5.1
blessings==1.6
bpython==0.16
curtsies==0.2.11
cycler==0.10.0
decorator==4.0.11
googler==1.0.1
greenlet==0.4.11
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
itunespy==1.5.3
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
matplotlib==2.0.0
nltk==3.2.1
numpy==1.11.3
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.1.3
pyparsing==2.1.10
pytaglib==1.4.0
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
requests==2.12.4
simplegeneric==0.8.1
six==1.10.0
tornado==4.4.2
traitlets==4.3.1
translate==2.0.0
wcwidth==0.1.7
wikipedia==1.4.0
